Set and change default env values.
The rule of thumb was values which are secret in production, shall not
be set via fallback value, to ensure failures if it was missed to set
them.
The .env.example has the pertaining variables enabled, so that the .env
file only has to be copied from it and the setup will work for local
development.